requirements.txt
# Monday Secretary dependencies
gspread==6.0.0
google-auth>=2.28
google-api-python-client>=2.122
notion-client>=2.2.1
pydantic>=2.7
python-dotenv>=1.0
requests>=2.31
tenacity>=8.2
pyyaml>=6.0
